Refactor Search input into `components/inputs` and hides native clear button
